Make hover (light move) link text visible
The hover text cooler is currently the same as the background colour; which means the text disappears on hover in light mode. I don't know what the best hover colour is, but this is where it needs to change in code. Feel free to close and implement in another PR.